Refactor WeatherInfo class
- Builder constructor takes now 3 args: city name, temp and temp
  unit. This is the minimun data that a weather service should
  provide when a weather update request is processed.
- Float members have been changed to double. Setter/getter methods
  updated.
- New setter/getter methods have been added to set current day high
  and low temp. Forecast list should be used only to provide weather
  forecast for upcoming days (this has been clearly documented).
  WeatherContract was updated to include these 2 new columns
- Added javadoc to all public methods
- Timestamp is not longer required in constructor. A new method
  setTimestamp has been added

TICKET: CYNGNOS-2365
TICKET: CYNGNOS-2382
TICKET: CYNGNOS-2356
TICKET: CYNGNOS-2360

+++ sdk/src/java/cyanogenmod/weather/util/WeatherUtils.java
@@ -31,8 +31,8 @@ public class WeatherUtils {
      * @param celsius temperature in Celsius
      * @return the temperature in degrees Fahrenheit
      */
-    public static float celsiusToFahrenheit(float celsius) {
-        return ((celsius * (9f/5f)) + 32f);
+    public static double celsiusToFahrenheit(double celsius) {
+        return ((celsius * (9d/5d)) + 32d);
     }
 
     /**
@@ -40,8 +40,8 @@ public class WeatherUtils {
      * @param fahrenheit temperature in Fahrenheit
      * @return the temperature in degrees Celsius
      */
-    public static float fahrenheitToCelsius(float fahrenheit) {
-        return  ((fahrenheit - 32f) * (5f/9f));
+    public static double fahrenheitToCelsius(double fahrenheit) {
+        return  ((fahrenheit - 32d) * (5d/9d));
     }
 
     /**
@@ -52,9 +52,9 @@ public class WeatherUtils {
      * @return A string with the format XX&deg;F or XX&deg;C (where XX is the temperature)
      * depending on the temperature unit that was provided or null if an invalid unit is supplied
      */
-    public static String formatTemperature(float temperature, int tempUnit) {
+    public static String formatTemperature(double temperature, int tempUnit) {
         if (!isValidTempUnit(tempUnit)) return null;
-        if (Float.isNaN(temperature)) return "-";
+        if (Double.isNaN(temperature)) return "-";
 
         DecimalFormat noDigitsFormat = new DecimalFormat("0");
         String noDigitsTemp = noDigitsFormat.format(temperature);
